Changed Copy to Extend
and made richError immutable.  With*** functions return a copy, storing the original as the underlying.

Getter methods all fall through to the underlying error.  This might make the error hierarchy a bit deeper, but makes Extend and Is work as expected in all cases, and ensures errors are never accidentally modified in place.