Add support for custom stack capture callback
This commit partially addresses issue #8.

When a merry error is first created, or first wrapped around an existing
error, a stacktrace is captured. That capturing is currently a hardcoded
internal function of merry. This commit turns the stacktrace function
into a callback. This allows users of merry to:

- remove the callback, to skip capturing stacktraces (for performance)
- replace the callback, to capture a different kind of stacktrace (e.g.
  sentry stacktraces)

The callback is registered globally. It is invoked whenever a new merry
error is created, or when an error was wrapped with a merry error. It is
passed the number of frames to skip, and returns []uintptr containing
the stack trace data.

Possible future enhancements:
Register a custom Stacktrace() function when registering the stack
capture function, so that the Stacktrace() method will continue to work
even with custom stacktraces.