Change the default behavior of Error()
Error() will fall back on the UserMessage if Message is empty.

Before, when you set the user message, if the message was empty *at that time*, we’d set both the user message and the message to the argument.  We no longer do that, in lieu of Error()’s new behavior.