fixes for errors with causes
Errors with a cause now use some special unwrap/is/as logic to ensure that only the latest cause attached to an error is traversed by Is and As.  Older causes are effectively overwritten and ignored.

Also, calling Unwrap in a loop on an error will now return the entire wrapper stack first, and then return the latest cause, so Unwrap can traverse the entire causal chain in the expected order (still ignoring overridden causes, as with As/Is)