Fixed an issue looking for values
Due to errWithCause's funky unwrap implementation, the simple scanning through the error stack for values wasn't working, because Unwrap wasn't returning raw *errImpls anymore.  So there is a new implementation of Value and Values which traverses wrapped errors, but not causes.

Also added Lookup() which is like Value, but returns a bool indicating whether the value was actually found or not.