Build the go-errors integration in natively, no hook required
Also prune back the grpc package to be a little less ambitious.  This package originally assumed grcp Statuses were meant to essentially be an encoding of an error, and that it would be useful to be able to convert both from errors to Statuses, and back again.  However, after using this package in a real context, it's clearer that Status is meant to be a one-way, lossy encoding of an error, and that there is little point converting back into an error.